Update patch set 1
Patch Set 1:

The reason phones limit the charging current when connected to USB ports is to comply with USB 2.0 specifications, which limit the supply current to 500 mA. I don't see how applying this patch would harm any hardware. This patch doesn't force your PC to supply more current to the USB port than the port could handle. Your devices are safe as long as the charging voltage is limited to 5.0 V, and this patch doesn't do anything to change that.

Patch-set: 1