Update patch set 1
Patch Set 1:

Aid - That's EXACTLY why this patch would harm hardware.  If you try to pull more than 500 mA from a port that was only designed to provide 500 mA, you're either going to fry/overheat that port, trip a fuse (fortunately most of these are breakers nowadays, but not all - I have a monitor with a dead USB hub to prove that), or brown out the device which can damage the device itself (one I777 user had that happen).

Daniel - The battery charging specification relies on some special detection logic to allow a device to detect when a high-current "charger" port is detected.  This patch causes high-current charging when a *normal* USB port is detected.

So far we're 2 for 2 on electrical engineers thinking this patch is bad, both a student and an old man.  :)

(Yes, I am an EE too.)

Patch-set: 1