Update patch set 1
Patch Set 1:

Glad I called in the EEs :)

It's easy to rationalize these kinds of changes away, but the reality is that the hardware we are dealing with is so varied that eventually someone is going to end up with a bomb on their desk. Remember those Evo guys who were "trickle charging" (overcharging) the batteries and causing them to catch on fire? And the crazy angry mob of XDA fanboys who didn't care? Let's keep to the spec :)

Patch-set: 1