Update patch set 1
Patch Set 1:

From an electrical standpoint this is not any more likely to harm the phone than charging via AC, but it may harm the other end of the chain, the USB-Host-Port and as such may render a USB-Port on a notebook unusable, thoug this is very unlikely it is still possible

however there are no serious dangers to be expected, since we already provide OC and other features that could harm a phone priced around $500 I see no point why we could not offer this feature too, as long as we provide the user with a defined warning which he has to accept.

Kernels like franco's have had this feature since ages and so far no harm done (seem them as betatesters) I myself use it a lot to charge my phone via my Macbook when I am on the go and I really appreciate the feature because pure USBcharging can hadly keep the charge when tethering on the go and transferring "some" data via HSDPA.

All in all I understand the outed concerns and would also consider a rework with a defined warning to the user.

Patch-set: 1