Update patch set 1
Patch Set 1:

Steve, you beat me to that comment regarding disabling fsync()...

I've put it in kernels before as an option with a huge THIS WILL SACRIFICE DATA INTEGRITY warning - but making it default AND impossible to disable is just wrong.

It would be nice if we start seeing mobile devices with USB switch chipsets that can detect/negotiate with CDPs.  (Actually, does the GNex use FSA9480 or TI's own TWL6040 for that function?)

Patch-set: 1